services/nlu-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
boto3==1.34.34
botocore==1.34.34
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.27.0
structlog==24.1.0
numpy==1.26.4
scikit-learn==1.4.1.post1
PyJWT[crypto]==2.8.0
aioboto3==12.3.0
cryptography==42.0.5
tenacity==8.2.3
prometheus-client==0.20.0
